Changes in package structure and documented files, also added license headers
