+++ nbproject/project.properties
@@ -5,6 +5,7 @@ annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
 application.title=ImageProcess
 application.vendor=Perin
+auxiliary.org-netbeans-spi-editor-hints-projects.perProjectHintSettingsFile=nbproject/cfg_hints.xml
 build.classes.dir=${build.dir}/classes
 build.classes.excludes=**/*.java,**/*.form
 # This directory is removed when the project is cleaned:
@@ -30,7 +31,9 @@ dist.javadoc.dir=${dist.dir}/javadoc
 endorsed.classpath=
 excludes=
 includes=**
+jar.archive.disabled=${jnlp.enabled}
 jar.compress=false
+jar.index=${jnlp.enabled}
 javac.classpath=
 # Space-separated list of extra javac options
 javac.compilerargs=
@@ -46,7 +49,7 @@ javac.test.classpath=\
 javac.test.processorpath=\
     ${javac.test.classpath}
 javadoc.additionalparam=
-javadoc.author=false
+javadoc.author=true
 javadoc.encoding=${source.encoding}
 javadoc.noindex=false
 javadoc.nonavbar=false
@@ -56,11 +59,29 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
+jnlp.codebase.type=no.codebase
+jnlp.descriptor=application
+jnlp.enabled=false
+jnlp.mixed.code=default
+jnlp.offline-allowed=false
+jnlp.signed=false
+jnlp.signing=
+jnlp.signing.alias=
+jnlp.signing.keystore=
 main.class=br.jpe.main.Main
+# Substitui\u00e7\u00e3o opcional do atributo padr\u00e3o Application-Library-Allowable-Codebase que identifica os locais onde se espera que o RIA assinado seja encontrado.
+manifest.custom.application.library.allowable.codebase=
+# Substitui\u00e7\u00e3o opcional do atributo padr\u00e3o Caller-Allowable-Codebase que identifica os dom\u00ednios dos quais o c\u00f3digo JavaScript pode fazer chamadas para o RIA sem avisos de seguran\u00e7a.
+manifest.custom.caller.allowable.codebase=
+# Substitui\u00e7\u00e3o opcional do atributo de manifesto da Base de c\u00f3digo; use para impedir que RIAs sejam adaptados
+manifest.custom.codebase=
+# Substitui\u00e7\u00e3o opcional do atributo de manifesto de Permiss\u00f5es (valores suportados: sandbox, all-permissions)
+manifest.custom.permissions=
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
 platform.active=default_platform
+project.license=gpl30
 run.classpath=\
     ${javac.classpath}:\
     ${build.classes.dir}
