+++ src/br/jpe/main/core/load/scripts/ColoredProcessScript.java
@@ -14,41 +14,24 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package br.jpe.main.core;
+package br.jpe.main.core.load.scripts;
+
+import java.awt.Color;
+import java.awt.image.BufferedImage;
+import br.jpe.main.core.load.PixelLoadScript;
 
 /**
- * An image bean
+ * A pixel script to load colored images
  *
  * @author joaovperin
  */
-public class Image {
+public class ColoredProcessScript implements PixelLoadScript {
 
-    private final double[][][] mtz;
-    private final int width;
-    private final int height;
-    private final int bands;
-
-    public Image(double[][][] mtz) {
-        this.mtz = mtz;
-        this.width = mtz.length;
-        this.height = mtz[0].length;
-        this.bands = mtz[0][0].length;
+    @Override
+    public void run(double[][][] mtz, BufferedImage img, int i, int j) {
+        Color color = new Color(img.getRGB(i, j));
+        mtz[i][j][0] = color.getRed();
+        mtz[i][j][1] = color.getGreen();
+        mtz[i][j][2] = color.getBlue();
     }
-
-    public double[][][] getMatrix() {
-        return mtz;
-    }
-
-    public int getWidth() {
-        return width;
-    }
-
-    public int getHeight() {
-        return height;
-    }
-
-    public int getBands() {
-        return bands;
-    }
-
 }
