+++ src/br/jpe/main/core/load/scripts/DecompositionProcessScript.java
@@ -1,0 +1,58 @@
+/*
+ * Copyright (C) 2018 Perin
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package br.jpe.main.core.load.scripts;
+
+import java.awt.Color;
+import java.awt.image.BufferedImage;
+import br.jpe.main.core.load.PixelLoadScript;
+
+/**
+ * A pixel script to load images as grayscale using decomposition method
+ *
+ * @author joaovperin
+ */
+public class DecompositionProcessScript extends DesaturationProcessScript implements PixelLoadScript {
+
+    private final boolean max;
+
+    private DecompositionProcessScript(boolean max) {
+        this.max = max;
+    }
+
+    public static final DecompositionProcessScript max() {
+        return new DecompositionProcessScript(true);
+    }
+
+    public static final DecompositionProcessScript min() {
+        return new DecompositionProcessScript(false);
+    }
+
+    @Override
+    public void run(double[][][] mtz, BufferedImage img, int i, int j) {
+        Color color = new Color(img.getRGB(i, j));
+
+        int value = (int) (max
+                ? maxValue(color.getRed(), color.getGreen(), color.getBlue())
+                : minValue(color.getRed(), color.getGreen(), color.getBlue()));
+
+        int numBands = mtz[0][0].length;
+        for (int n = 0; n < numBands; n++) {
+            mtz[i][j][n] = value;
+        }
+    }
+
+}
