+++ src/br/jpe/main/core/load/scripts/SingleColorProcessScript.java
@@ -1,19 +1,31 @@
 /*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
+ * Copyright (C) 2018 Perin
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package br.jpe.main.core.scripts;
+package br.jpe.main.core.load.scripts;
 
-import br.jpe.main.core.PixelScript;
 import java.awt.Color;
 import java.awt.image.BufferedImage;
+import br.jpe.main.core.load.PixelLoadScript;
 
 /**
+ * A pixel script to load images as grayscale using single color picking method
  *
- * @author Perin
+ * @author joaovperin
  */
-public class SingleColorProcessScript extends DesaturationProcessScript implements PixelScript {
+public class SingleColorProcessScript extends DesaturationProcessScript implements PixelLoadScript {
 
     private static final int CHANNEL_RED = 1;
     private static final int CHANNEL_GREEN = 2;
